added perfect number program
Any number can be perfect if the sum of its divisors (excluding the actual number) equal to the number. For instance, 6 is divisible by 1, 2, 3 (excluding 6) and the sum of 1 + 2 + 3 = 6. So, 6 is a perfect number.

In this C++ perfect number example, we used for loop (for(i = 1 ; i < number ; i++)) to iterate from 1 to number. We used the if condition (if(number % i == 0)) to check whether the number divisible by loop value equals 0. If true, add that value to the sum. Finally, we used the If condition (if(number == sum)) to check whether the number equal to the sum and print the result.